Rename clear button component
RISDEV-8269
