+++ frontend/src/components/WithClearButton.vue
